+++ cmd/entrello/main.go
@@ -6,13 +6,21 @@ import (
 	"time"
 
 	"github.com/utkuufuk/entrello/internal/config"
+	"github.com/utkuufuk/entrello/internal/github"
 	"github.com/utkuufuk/entrello/internal/syslog"
+	"github.com/utkuufuk/entrello/internal/tododock"
 	"github.com/utkuufuk/entrello/internal/trello"
 )
 
 var (
 	logger syslog.Logger
 )
+
+type CardQueue struct {
+	add chan trello.Card
+	del chan trello.Card
+	err chan error
+}
 
 func main() {
 	// read config params
@@ -49,7 +57,54 @@ func main() {
 	// concurrently fetch new cards from sources and start processing cards to be created & deleted
 	q := CardQueue{make(chan trello.Card), make(chan trello.Card), make(chan error)}
 	for _, src := range sources {
-		go queueActionables(ctx, src, client, q)
+		go src.queueActionables(ctx, client, q)
 	}
 	processActionables(ctx, client, q)
+}
+
+// getEnabledSourcesAndLabels returns a slice of enabled sources & their labels as a separate slice
+func getEnabledSourcesAndLabels(cfg config.Sources) (sources []source, labels []string) {
+	arr := []source{
+		{cfg.GithubIssues.SourceConfig, github.GetSource(cfg.GithubIssues)},
+		{cfg.TodoDock.SourceConfig, tododock.GetSource(cfg.TodoDock)},
+	}
+
+	now := time.Now()
+
+	for _, src := range arr {
+		if ok, err := src.shouldQuery(now); !ok {
+			if err != nil {
+				logger.Errorf("could not check if '%s' should be queried or not, skipping", src.cfg.Name)
+			}
+			continue
+		}
+		sources = append(sources, src)
+		labels = append(labels, src.cfg.Label)
+	}
+	return sources, labels
+}
+
+// processActionables listens to the card queue in an infinite loop and creates/deletes Trello cards
+// depending on which channel the cards come from. Terminates whenever the global timeout is reached.
+func processActionables(ctx context.Context, client trello.Client, q CardQueue) {
+	for {
+		select {
+		case c := <-q.add:
+			if err := client.CreateCard(c); err != nil {
+				logger.Errorf("could not create Trello card: %v", err)
+				break
+			}
+			logger.Printf("created new card: %s", c.Name)
+		case c := <-q.del:
+			if err := client.ArchiveCard(c); err != nil {
+				logger.Errorf("could not archive card card: %v", err)
+				break
+			}
+			logger.Printf("archived stale card: %s", c.Name)
+		case err := <-q.err:
+			logger.Errorf("%v", err)
+		case <-ctx.Done():
+			return
+		}
+	}
 }
